pyproject.toml
[project]
name = "django-boilerplate"
version = "0.1.0"
description = "Django Boilerplate Project"
authors = [
    {name = "Sujit Dhakal"}
]
requires-python = "^3.12"
dependencies = [
    "django (>=5.0,<6.0)",
    "python-decouple (>=3.8,<4.0)",
    "celery (>=5.6.2,<6.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "django-filter (>=25.2,<26.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "drf-yasg (>=1.21.15,<2.0.0)",
    "channels-redis (>=4.3.0,<5.0.0)",
    "redis (>=7.2.1,<8.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "gunicorn (>=25.1.0,<26.0.0)",
    "typing-extensions (>=4.15.0,<5.0.0)",
    "pandas (>=3.0.1,<4.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
